+++ src/About/DeviceCard.tsx
@@ -32,7 +32,7 @@ export default ({ className }: DeviceCardProps) => {
     if (device == null) {
         return (
             <Card className={className}>
-                <Card.Header className="tw-text-center">
+                <Card.Header className="tw-text-center tw-text-base">
                     <Card.Header.Title cardTitle="Device" />
                 </Card.Header>
                 <Card.Body>
@@ -47,7 +47,7 @@ export default ({ className }: DeviceCardProps) => {
 
     return (
         <Card className={className}>
-            <Card.Header className="tw-text-center">
+            <Card.Header className="tw-text-center tw-text-base">
                 <Card.Header.Title cardTitle="Device" />
             </Card.Header>
             <Card.Body className="tw-gap-4">