+++ src/About/DocumentationCard.tsx
@@ -18,7 +18,7 @@ export default ({
     className,
 }: DocumentationCardProps) => (
     <Card className={className}>
-        <Card.Header className="tw-text-center">
+        <Card.Header className="tw-text-center tw-text-base">
             <Card.Header.Title cardTitle="Documentation" />
         </Card.Header>
         <Card.Body className="tw-gap-4">{documentationSections}</Card.Body>